+++ webdev/test/serve/injected/devtools_test.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @Timeout(Duration(minutes: 5))
-@Tags(['webdriver'])
 import 'dart:async';
 import 'dart:io';